Changes to be committed: 	modified:   Renderer.class 	modified:   Renderer.java 	modified:   Rendering$1$1.class 	new file:   Rendering$1$10.class 	modified:   Rendering$1$2.class 	modified:   Rendering$1$3.class 	modified:   Rendering$1$4.class 	modified:   Rendering$1$5.class 	modified:   Rendering$1$6.class 	modified:   Rendering$1$7.class 	new file:   Rendering$1$8.class 	new file:   Rendering$1$9.class 	modified:   Rendering$1.class 	modified:   Rendering.class 	modified:   Rendering.java

+++ Renderer.java
@@ -1,16 +1,17 @@
 import java.util.List;
 import java.awt.image.BufferedImage;
+import java.awt.image.DataBufferInt;
 import java.awt.Color;
 
 public class Renderer {
-    private static final double brightness = 10;
+    private static final float brightness = 10f;
 
     public BufferedImage render(List<Triangle> tris, Matrix3 transform, int width, int height, boolean[] highlight, int playerX, int playerZ) {
         BufferedImage img = new BufferedImage(width, height, BufferedImage.TYPE_INT_ARGB);
-        double[] zBuffer = new double[img.getWidth() * img.getHeight()];
+        float[] zBuffer = new float[img.getWidth() * img.getHeight()];
 
         for (int i = 0; i < zBuffer.length; i++) {
-            zBuffer[i] = Double.NEGATIVE_INFINITY;
+            zBuffer[i] = Float.NEGATIVE_INFINITY;
         }
 
         for (int i = 0; i < tris.size(); i++) {
@@ -26,7 +27,7 @@ public class Renderer {
         Triangle t, 
         Matrix3 transform, 
         BufferedImage img, 
-        double[] zBuffer,
+        float[] zBuffer,
         Color useColor,
         int playerX,
         int playerZ
@@ -81,35 +82,35 @@ public class Renderer {
         );
 
         // Normalize
-        double len = Math.sqrt(wnorm.getX()*wnorm.getX() + wnorm.getY()*wnorm.getY() + wnorm.getZ()*wnorm.getZ());
+        float len = (float) Math.sqrt(wnorm.getX()*wnorm.getX() + wnorm.getY()*wnorm.getY() + wnorm.getZ()*wnorm.getZ());
         wnorm.setX(wnorm.getX() / len);
         wnorm.setY(wnorm.getY() / len);
         wnorm.setZ(wnorm.getZ() / len);
 
         // Directional light (sun)
         Vertex lightDir = new Vertex(0.3, 1.0, 0.2); // slightly angled sunlight
-        double lLen = Math.sqrt(lightDir.getX()*lightDir.getX() + lightDir.getY()*lightDir.getY() + lightDir.getZ()*lightDir.getZ());
+        float lLen = (float) Math.sqrt(lightDir.getX()*lightDir.getX() + lightDir.getY()*lightDir.getY() + lightDir.getZ()*lightDir.getZ());
         lightDir.setX(lightDir.getX() / lLen);
         lightDir.setY(lightDir.getY() / lLen);
         lightDir.setZ(lightDir.getZ() / lLen);
 
         // Dot product for brightness
-        double angleCos = Math.max(0.2, wnorm.getX()*lightDir.getX() + wnorm.getY()*lightDir.getY() + wnorm.getZ()*lightDir.getZ());
+        float angleCos = (float) Math.max(0.2, wnorm.getX()*lightDir.getX() + wnorm.getY()*lightDir.getY() + wnorm.getZ()*lightDir.getZ());
 
         int minX = (int) Math.max(0, Math.ceil(Math.min(v1.getX(), Math.min(v2.getX(), v3.getX()))));
         int maxX = (int) Math.min(width - 1, Math.floor(Math.max(v1.getX(), Math.max(v2.getX(), v3.getX()))));
         int minY = (int) Math.max(0, Math.ceil(Math.min(v1.getY(), Math.min(v2.getY(), v3.getY()))));
         int maxY = (int) Math.min(height - 1, Math.floor(Math.max(v1.getY(), Math.max(v2.getY(), v3.getY()))));
 
-        double triangleArea = (v1.getY() - v3.getY()) * (v2.getX() - v3.getX()) + (v2.getY() - v3.getY()) * (v3.getX() - v1.getX());
+        float triangleArea = (float) ((v1.getY() - v3.getY()) * (v2.getX() - v3.getX()) + (v2.getY() - v3.getY()) * (v3.getX() - v1.getX()));
 
         for (int y = minY; y <= maxY; y++) {
             for (int x = minX; x <= maxX; x++) {
-                double b1 = ((y - v3.getY()) * (v2.getX() - v3.getX()) + (v2.getY() - v3.getY()) * (v3.getX() - x)) / triangleArea;
-                double b2 = ((y - v1.getY()) * (v3.getX() - v1.getX()) + (v3.getY() - v1.getY()) * (v1.getX() - x)) / triangleArea;
-                double b3 = ((y - v2.getY()) * (v1.getX() - v2.getX()) + (v1.getY() - v2.getY()) * (v2.getX() - x)) / triangleArea;
+                float b1 = (float) ((y - v3.getY()) * (v2.getX() - v3.getX()) + (v2.getY() - v3.getY()) * (v3.getX() - x)) / triangleArea;
+                float b2 = (float) ((y - v1.getY()) * (v3.getX() - v1.getX()) + (v3.getY() - v1.getY()) * (v1.getX() - x)) / triangleArea;
+                float b3 = (float) ((y - v2.getY()) * (v1.getX() - v2.getX()) + (v1.getY() - v2.getY()) * (v2.getX() - x)) / triangleArea;
                 if (b1 >= 0 && b1 <= 1 && b2 >= 0 && b2 <= 1 && b3 >= 0 && b3 <= 1) {
-                    double depth = b1 * v1.getZ() + b2 * v2.getZ() + b3 * v3.getZ();
+                    float depth = (float) (b1 * v1.getZ() + b2 * v2.getZ() + b3 * v3.getZ());
                     int zIndex = y * width + x;
                     if (zBuffer[zIndex] < depth) {
                         img.setRGB(x, y, getShade(useColor, angleCos).getRGB());
@@ -120,12 +121,12 @@ public class Renderer {
         }
     }
 
-    private static Color getShade(Color color, double shade) {
+    private static Color getShade(Color color, float shade) {
         shade *= brightness;
 
-        double redLinear = Math.pow(color.getRed(), 2.4) * shade;
-        double greenLinear = Math.pow(color.getGreen(), 2.4) * shade;
-        double blueLinear = Math.pow(color.getBlue(), 2.4) * shade;
+        float redLinear = (float) Math.pow(color.getRed(), 2.4) * shade;
+        float greenLinear = (float) Math.pow(color.getGreen(), 2.4) * shade;
+        float blueLinear = (float) Math.pow(color.getBlue(), 2.4) * shade;
 
         int red = (int) Math.min(Math.pow(redLinear, 1/2.4), 255);
         int green = (int) Math.min(Math.pow(greenLinear, 1/2.4), 255);
